Refactor TypingAnimation component: rename file and update import path
