+++ src/app/Components/Pages/Aboutus.jsx
@@ -1,8 +1,9 @@
+"use client"
 import React from "react";
-import TypingAnimation from "../../animation/typing-animation.tsx";
+import TypingAnimation from "../../animation/typinganimation";
 
 export default function Aboutus() {
-    
+
   return (
     <div className="py-[3%] max-w-screen-xl mx-auto">
       <div className="text-3xl text-blue-600 font-bold">
