+++ src/app/animation/typinganimation.tsx
